Move core vitals init to a component
